// backend/src/ai/schema-meta.generated.ts
export interface GeneratedField {
  name: string;
  type: string;
  isList: boolean;
  isOptional: boolean;
  default?: string;
  enumValues?: string[];
  relationTo?: string;
  comment?: string;
}